use a secure temporary directory
the previous code was using a fixed-name for a temporary file that was
used to store the name of the file to be uploaded. an attacker could
abuse this to make teiler upload arbitrary files the user has access
to to a paste site. since most paste sites supported are fairly
public, this could lead to elevated privileges, information disclosure
or even remote code execution (think of .ssh/id_rsa).